+++ src/helpers/storage/removeStorage.js
@@ -1,12 +1,18 @@
 /**
  * Removes a value from localStorage and its associated cookie.
+ *
+ * @module helpers/storage/removeStorage
+ *
  * @param {string} key - Storage key.
  */
+
+import removeCookie from './removeCookie' ;
+
 const removeStorage = key =>
 {
     localStorage.removeItem( key ) ;
 
-    document.cookie = `${key}=;path=/;max-age=0` ;
+    removeCookie( key ) ;
 
     window.dispatchEvent( new StorageEvent( 'storage' , { key } ) ) ;
 } ;